+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 NEXT
 ====================
+- Improved `Apify.launchPuppeteer()` code to prevent triggering bugs in Puppeteer by passing
+  more than required options to `puppeteer.launch()`.
 
 0.19.1 / 2020-01-30
 ====================